Edit fabs(int) in GSL codes to avoid clang warning

+++ src/gsl/specfunc/poch.c
@@ -166,7 +166,7 @@ pochrel_smallx(const double a, const double x, gsl_sf_result * result)
 
     if(bp == a) {
       result->val = dpoch1;
-      result->err = 2.0 * GSL_DBL_EPSILON * (fabs(incr) + 1.0) * fabs(result->val);
+      result->err = 2.0 * GSL_DBL_EPSILON * (fabs((double)(incr)) + 1.0) * fabs(result->val); // edited for qfratio
       return GSL_SUCCESS;
     }
     else {
@@ -181,7 +181,7 @@ pochrel_smallx(const double a, const double x, gsl_sf_result * result)
       double trig  = t1 - t2;
       result->val  = dpoch1 * (1.0 + x*trig) + trig;
       result->err  = (fabs(dpoch1*x) + 1.0) * GSL_DBL_EPSILON * (fabs(t1) + fabs(t2));
-      result->err += 2.0 * GSL_DBL_EPSILON * (fabs(incr) + 1.0) * fabs(result->val);
+      result->err += 2.0 * GSL_DBL_EPSILON * (fabs((double)(incr)) + 1.0) * fabs(result->val); // edited for qfratio
       return GSL_SUCCESS;
     }    
   }
